Refactoring of monitoring_window_config
